create ft_strjoin.c & modify ft_strlen.c

+++ libft/ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	len;
